Remove MessageService
