+++ src/app/admin/admin.component.spec.ts
@@ -9,7 +9,6 @@ import { CollectionService } from 'app/core/collection-module/collections.servic
 import { ActivatedRoute, Router, NavigationEnd, Scroll } from '@angular/router';
 import { of } from 'rxjs';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { MessagesService } from 'app/core/utility-module/messages.service';
 
 describe('AdminComponent', () => {
   let component: AdminComponent;
@@ -40,7 +39,6 @@ describe('AdminComponent', () => {
         AuthService,
         ToastrOvenService,
         CollectionService,
-        MessagesService,
         { provide: Router, useValue: routerStub },
         { provide: ActivatedRoute, useClass: ActivatedRouteStub },
       ],
