+++ src/app/admin/admin.component.ts
@@ -6,7 +6,6 @@ import { takeUntil, skipWhile, take, filter, map, switchMap } from 'rxjs/operato
 import { AuthService } from 'app/core/auth-module/auth.service';
 import { CollectionService, Collection } from 'app/core/collection-module/collections.service';
 import { ToastrOvenService } from 'app/shared/modules/toaster/notification.service';
-import { MessagesService } from 'app/core/utility-module/messages.service';
 import { UtilityService } from 'app/core/utility-module/utility.service';
